requirements.txt
asyncpg==0.29.0
cryptography==41.0.5
passlib==1.7.4
pydantic==2.4.2
email-validator==2.1.0
pydantic-settings==2.0.3
aioredis==2.0.1
fastapi==0.103.2
uvicorn==0.23.2
gunicorn==21.2.0
orjson==3.9.9
python-dotenv==1.0.0
python-jose==3.3.0
sqlalchemy==2.0.23
greenlet==3.0.1
alembic==1.12.1
typer==0.9.0
